Adding changes required for this code to be used in IDC Inflow
In IDC Inflow images are provisioned using Ansible.
Since Ansible is not good with reboots performed within the script
and starting services doesn't make sence when building image a
few environment variables are introduced to block these actions.

OMMIT_SERVICE_START - when non-empty services are not started
OMMIT_REBOOT - wehn non-empty reboots are not performed

+++ base/5_env_dev_utils_setup.sh
@@ -59,7 +59,7 @@ EOF
 
 # enable and start the service
 sudo systemctl enable set-cpufreq-governor.service
-sudo systemctl start set-cpufreq-governor.service
+[ -z "$OMMIT_SERVICE_START" ] && sudo systemctl start set-cpufreq-governor.service
 
 # install required packages
 colored_output "Installing required packages..." blue
@@ -108,6 +108,10 @@ echo "ulimit -u 65535" | sudo tee -a /etc/security/limits.conf
 # inform user
 colored_output "Cleanup..." blue
 sudo apt -y autoremove
-colored_output "Setup completed. Rebooting in 10 seconds..." blue
-sleep 10
-sudo reboot
+
+if [ -z "$OMMIT_REBOOT" ]
+then
+    colored_output "Setup completed. Rebooting in 10 seconds..." blue
+    sleep 10
+    sudo reboot
+fi
